Fix RHEL 9.1 RPM build due to headers_intall changes
The script headers_install.sh has changed upstream. It now no longer take a
directory of header files to operate on instead it takes direct path to a signle
input and output file. The order of which is also reversed. Fix up our build
scripts to do the right thing on RHEL 9.1 and later.

Signed-off-by: Dennis Dalessandro <dennis.dalessandro@cornelisnetworks.com>