INTERNAL_SNOOP: Add OPXS snoop/capture functionality
Add diag.c from ifs-all opa-dev branch and necessary tweaks to get it to
compile.

Updates for RHEL 8.8:
--------------------
RedHat has included a patch that removed leftover diagpkt stuff. It needs added
back in for snoop/capture.

> commit b90c7e97c48bb1a94cd49cc32a4d2a62a06cbf1c
> Author: Dennis Dalessandro <dennis.dalessandro@cornelisnetworks.com>
> Date:   Fri May 20 14:37:27 2022 -0400
>
>    RDMA/hfi1: Remove all traces of diagpkt support
>
>    One of the concessions we made to get our driver upstream was to remove
>    the diagnostic packet support. There is however still some cruft that was
>    left over.  Remove it.
>
>    Link: https://lore.kernel.org/r/20220520183727.48973.93587.stgit@awfm-01.cornelisnetworks.com
>    Signed-off-by: Dennis Dalessandro <dennis.dalessandro@cornelisnetworks.com>
>    Signed-off-by: Jason Gunthorpe <jgg@nvidia.com>

Updates for RHEL 9.2
--------------------
While 9.1 was able to use the 8.7 version of snoop, we need the 8.8 version for
9.2 which is expected because the kernels are at the same level (RDMA wise).

The new compiler in 9.2 though needs us to add "fallthrough;" statements instead
of the /*Fall through*/ annotations that we used before so add those.

Updates for SLES 15.5
---------------------
Was able to use the RHEL 9.2 version but ended up with a double
definition of diagpkt in common.h, fixed that up and we are fine. SLES
needs the fall through updates, but not the diagpkt put back in thing.
This is the easiest change, just remove the extra definition.

Signed-off-by: Dennis Dalessandro <dennis.dalessandro@cornelisnetworks.com>
(cherry picked from commit 2ed53a2f9789f98d126be8202ba491eef5ba197e)
Signed-off-by: Dennis Dalessandro <dennis.dalessandro@cornelisnetworks.com>