Remove localbuild
Remove the top level makefile. No need to have 2 build systems. Running the
do-update-makerpms.sh script can be just as fast.

Here is an example of what I do to build on a local disk vs out of NFS:

 sdir=$PWD; rm -rf /tmp/tmpbuild && ./do-update-makerpm.sh -S ${PWD} \
 -w /tmp/tmpbuild && cd /tmp/tmpbuild/rpmbuild && rpmbuild --rebuild \
 --define  "_topdir $(pwd)" --nodeps SRPMS/*.src.rpm; cd $sdir

This builds the RPM in /tmp and then comes back to this directory.

Signed-off-by: Dennis Dalessandro <dennis.dalessandro@cornelisnetworks.com>