SLES: Disable CONFIG_FAULT_INJECTION
RHEL does not enable CONFIG_FAULT_INJECTION by default but SLES does. We do not
want this turned on in normal circumstances. It is a diagnostic/debugging option
that could impact performance to an unknown degree. Therefore we want it
disabled. Add the #undef macro to turn it off explicitly on SLES. We do this in
previous version of IFS/OPXS drivers by turning it off in the compat stuff. We
do not have that here so put it in fault.c.

Signed-off-by: Dennis Dalessandro <dennis.dalessandro@cornelisnetworks.com>