RDMA/hfi1: Setup VF/PF0 resources when SRIOV is configured
A VF acquires its resource limits from PF0 early in initialization, and
then uses those limits to initialize driver resource structures. PF0
also acquires SRIOV resources since they are limited compared to the
non-SRIOV case. VF drivers use a separate call to cause the resources to
be assigned to them (via SiIdx CSRs).

For testing purposes, SRIOV configuration is specified by module
parameters on the PF0 driver. The following are defined:

	max_num_vfs - maximum number of VFs allowed
	ctxt_per_vf - send/recv contexts per VF
	sdma_per_vf - SDMA engiones per VF
	rcv_per_vf  - RcvArray entries per VF
	pio_per_vf  - PIO memory blocks per VF

PF0 receives the remaining resources after the VFs. Resources are
partitioned according to 'max_num_vfs', not the actual number of VFs
created (<= max_num_vfs). PF0 resources are at the beginning of the
range, VF1 is at the end of the range. If no *_per_vf resources are
specified, defaults will be chosen based on reserving a minimal number
of resources for PF0.

The VF driver running in a VM must specify si_idx as the number related
to the PCI virtual function number of the device, with '1' being the
first PCI VF device.

Signed-off-by: Douglas Miller <doug.miller@cornelisnetworks.com>