RDMA/hfi1: Implement vf2pf-over-loopback
Reserves 8 contexts regardless of max_num_vfs. Contexts are reserved
from the top of the range, lowest number is for PF0, next for VF1, etc.

Only the PING message is implemented. A ping is initiated by writing
text to the VF's "vf2pf_ping" attribute, where a ping message containing
the text is sent to PF0 and echoed back to the VF with a '!' appended.
Messages are printed on dmesg, at both PF0 and the VF, to indicate the
progress of the ping.

Messages is sent using 9B KDETH packets, using PIO to send and receiving
as eager. Receive interrupts are used, although early initialization
requires a polling mode.

The VFs will attempt to deterine their SI number if it is not provided
with the si_idx module parameter, by probing the RcvHdrEntSize CSR and
looking for a non-zero value, which is only seen of the VF was assigned
the receive context.

The "hw_resopurces" sysfs attribute is extended to include context
resources assign for vf2pf. In addition, there is a "vf2pf_lb_debug"
attribute that shows more detailed information on the resources.

Signed-off-by: Douglas Miller <doug.miller@cornelisnetworks.com>