BACKPORT: RHEL 9.4 compatibility patch
The only difference between RHEL9.4 and 9.5 distro drivers is a
vm_flags_clear/reset calls in file_ops.c. Touch that up and we should have
a RHEL 9.4 build working.

Well almost anyway.

We need to add a BIT() macro

9.4 DEFINE_SEMAPHORE macro defaults to a 1, instead of taking an argument.
We may need to start off with a down() on the sem because it was getting
inited to 0 in our upstream/9.5 code.

Signed-off-by: Dennis Dalessandro <dennis.dalessandro@cornelisnetworks.com>