INTERNAL: reference correct Nvidia source package
Driver builds depend on the Nvidia source package, not the binary
package. It turns out that these can vary based on options chosen during
CUDA install, and the version was being extracted incorrectly in some
cases.

Change the build script and the hfi1 Makefile to reference the correct
package build builds and .deb dependencies. Additionally, remove an
unnecessary conditional from the Makefile that is only needed for Red
Hat builds.

Signed-off-by: Mitch Williams <mitch.williams@cornelisnetworks.com>