RDMA/hfi1: Initialize loopback ports and use when needed
In order to use loopback ports it is required that loopback port-specific
CSRs be initialized. This code will initialize loopback CSRs identically
to their corresponding fabric port CSRs.

The relationship of the loopback port index to the fabric port index is
contained entirely in loopback_pidx_dd(), and by proxy loopback_pidx().
This is only a convention, there is no physical requirement that a
particular loopback port be paired with a particular fabric port.

Detect situations where SLID == DLID but are different VFs (or PF0).
In those cases, rvt_ruc_loopback() or similar tricks cannot be used.
Use the loopback port but otherwise send normally. A loopback port
is implicitly associated with a fabric port but does not have a separate
struct hfi1_pportdata. The selection of loopback vs. fabric port is
done at the point that the PBC is created. JKR cannot use the loopback
port to avoid receiving from more than one port.

TODO: ipoib (not working)

Signed-off-by: Douglas Miller <doug.miller@cornelisnetworks.com>