RDMA/hfi1: Handle RcvQPMapTable setup for VFs
Change the 24-bit QP number space such that bits 8:1 are limited to values
in the range of receive contexts assigned to the VF.  This ensures that
the QPN is unique for the adapter and that the RcvQPMapTable entries are
private to the VF. A QPN might arrive at either port, so the RcvQPMapTable
needs to be programmed for the full range on both ports.

The reason this is not done once by PF0 for all is that it is possible
that a VF might have a different number of kernel contexts (based on
configuration that PF0 does not know about). A VF passes information
about kernel contexts to PF0 during its init_rxe() sequence, allowing
for the RcvQPMapTable entries to be customized for the VF.

Also, restore the PF0-only RcvQPMapTable when SRIOV is disabled.

Signed-off-by: Douglas Miller <doug.miller@cornelisnetworks.com>