RDMA/hfi1: Check for the SCF_LINK_DOWN flag in sc_wait_for_packet_egress
When waiting to egress a packet out of a send context, check
for if the send context has the SCF_LINK_DOWN flag set before
considering whether to throw and handle an egress timeout
error. This SCF_LINK_DOWN flag will be set from start_linkdown_handling
when hfi1 handles a linkdown event.

Signed-off-by: Manuel Gomez <manuel.gomez@cornelisnetworks.com>
Signed-off-by: Dennis Dalessandro <dennis.dalessandro@cornelisnetworks.com>