RDMA/hfi1: Support additional MADs through MCTXT UMAD channel
Prevent reflection of CPORT MADs that target local node: As CPORT starts
to send more types of MADs over the MCTXT UMAD path, those MADs must
not be reflected back to the CPORT. Use the 'vendor_err' field in the
ib_wc to tag the MAD between cport_umad_handler() and cport_process_mad()
so that we know not to reflect it back.

Use the correct SLID, DLID, and PKEY values when building the struct
ib_wc for passing to rvt_recv_cq(). The original use case, firmware
update, did not require special handling of these values.

Requires that firmware be updated to at least 12_0_2_0_3 before
updating driver.

<Signed-off-by: Douglas Miller <doug.miller@cornelisnetworks.com>
<Signed-off-by: Dennis Dalessandro <dennis.dalessandro@cornelisnetworks.com>