hfi1/user_pages: Don't set locked for remote pages
Bulksvc cannot pin any pages because of old kernel rule where
the locked arg cannot be non-NULL with FOLL_LONGTERM flag set.

This dependency is summarized and fixed in a patch in v6.2 [2]

Since the locked flag seems to only be to allow the invoked function
to unlock to recover from VM_FAULT_RETRY, just don't give that
permission.

[1] - https://github.com/torvalds/linux/blob/7d2a07b769330c34b4deabeed939325c77a7ec2f/mm/gup.c#L1849-L1866
[2] - https://github.com/cornelisnetworks/wfr-linux-devel/commit/53b2d09bdd12092a7341c08b6b863560db62fa57

Signed-off-by: Nick Child <nchild@cornelisnetworks.com>
Signed-off-by: Dennis Dalessandro <dennis.dalessandro@cornelisnetworks.com>