RDMA/hfi1: Allow setting of PCIe completion timeout
Add a module parameter to specify the PCIe capabilities setting
for completion timeout. Use a default valule of 0x06 (ranges B, C)
and allow for also setting the disable bit (0x10) to disable
completion timeout. Specifying a value of 0 results in no change,
however a restart of firmware and BIOS may be required to restore
a platform default.

INTERNAL: STL-75112

Signed-off-by: Douglas Miller <doug.miller@cornelisnetworks.com>
Signed-off-by: Dennis Dalessandro <dennis.dalessandro@cornelisnetworks.com>