bulksvc: add strictly ordered data transfers
Enhancement to the hfi1 driver dms component to provide strictly
ordered data transfers in support of the verbs ordering semantics.

The unordered data transfer protocol is simplified and unified with the
new ordered data transfer protocol. As such the "small" and "fixup"
(unaligned) data packet types have been eliminated. The alignment fixup
head/tail data is now included in the head of the data packet type.

A new ordering concept, the "ordering domain", is introduced which
manages the queues and counters associated with data flows which must
retain order through completion with respect to each other. The ordering
domain will "pin" hardware resources, such as sdma engines, so that all
transfers active on a domain will be ensured to remain in order through
the hardware. The order domain also manages access to hardware, such as
tidsets and rift entries, so that data flows on a single order domain do
not starve other order domains of resources.

All read/write combinations are supported (RAR, WAW, RAW, WAR) by using
various domain transfer state counters to determine if a transfer
protocol may issue a hardware sdma operation. For the RAW and WAR cases
the protocol must delay until a previous operation has completed
receiving. The protocols are written such that if all transfers types
are the same then no delays will be required because the sdma hardware
can enforce the ordering.

INTERNAL: STL-75667

Signed-off-by: Michael Blocksome <michael.blocksome@cornelisnetworks.com>
Signed-off-by: Dennis Dalessandro <dennis.dalessandro@cornelisnetworks.com>