RDMA/hfi1: Receive over-temp alert from CPORT
Register for the new MCTXT trap generated by firmware when over-temp
conditions are reached, at which point shutdown of the hardware is
imminent (500mS). Print a message indicating the cause of the impending
hangs/panic.

The driver is not designed to take any further action, such as
unbinding the device or throttling I/O.

INTERNAL: STL-75190

Signed-off-by: Douglas Miller <doug.miller@cornelisnetworks.com>
Signed-off-by: Dennis Dalessandro <dennis.dalessandro@cornelisnetworks.com>