INTERNAL: STL-75236 fix Ubuntu build dependency issues
With the existing .deb file naming convention, it was possible for the
user to install both the non-gpu and gpu modules together, leading to
messed-up results and broken packages. To fix this, we change the
package name to include the -rocm or -cuda suffix, instead of including
it in the version number as before. This allows us to then designate
these packages as conflicting so that you cannot install the base
package alongside a gpu package.

Along with this, we simplify the version numbering to just specify the
base name of the kernel, instead of the full patch id, because DKMS
allows us to install on (hopefully) any Ubuntu patch version of the same
kernel. To designate which portion of the version is our revision id,
add an "r" to version string to separate this.

Finally, we now build a -dev package to update headers for userspace.
This has the same revision name as the dkms package and depends on the
dkms package for installation. Although the headers are currently
identical, we plan for the future by building a separate -dev package
for each GPU variant.

Signed-off-by: Mitch Williams <mitch.williams@cornelisnetworks.com>