RDMA/hfi1: Process MCTXT messages larger than 1 packet
Previously, it was assumed all MCTXT messages were less than the MCTXT
packet window of 2048 bytes.
This meant that EOM and SOM were assumed to always be set. There was no
mechanism to sent/receive messages that exceeded 2048 bytes.

To handle >2048 the rsp and req buffers must be dynamically allocated.
For consistency, allocate in units of 2048 bytes and since request
size does not imply response size we must have a way to reallocate when
the response comes back.

It seems in 99% of circumstances only 1 mctxt packet will be needed for
both request and response so optimize towards that.

When sending to cport, we can only send the first 2048 bits of the
packet. Since the sending request comes in on a workqueue, we simply
bump the message data pointer and reschedule the work. There may be
some serialization work left to implement here.

When receiving from cport, we won't have a reference to the current
message so create a pointer to cport to track an expected incomplete
message. This is safe since cport receive is single threaded.

INTERNAL: STL-75446

Signed-off-by: Nick Child <nchild@cornelisnetworks.com>
Signed-off-by: Dennis Dalessandro <dennis.dalessandro@cornelisnetworks.com>