RDMA/hfi1: Complete SDMA transmits when an error occurs
Verbs transmits cannot be just dropped - they must be completed.  When
a pending SDMA transmit fails to queue or any other unexpected error
occurs, call the complete function with an error.

In the SDMA complete function, update the owning QP with a result that
reflects the status passed in.  For RC QPs, recognize that any error
will result in an RC abort, not just an internal SDMA engine error.

If a txreq build fails, queueing it as-is to try later is incorrect -
the txreq was not completely built.  If it were to be queued, the txreq
would need to be cleaned first.  Instead of queueing, complete the txreq
with an error.  The error that caused the failure will likely occur again.

INTERNAL: STL-75458 A verbs SDMA txreq is not completed if descriptors are full and the QP is in a bad state

Signed-off-by: Dean Luick <dean.luick@cornelisnetworks.com>
Signed-off-by: Dennis Dalessandro <dennis.dalessandro@cornelisnetworks.com>