RDMA/hfi1: Harden CPORT MCTXT state recovery
Detect MCTXT PF0 CSRs in a bad state and attempt to repair.  This is
controlled by module param 'cport_mctxt_recovery' which defaults to
'true'.

Use a timeout on the initial PING to avoid hangs trying to
flush the WQ on START failure. Add a module param 'cport_adm_to' to
allow configuration of the timeout used for all administrative (non-MAD)
CPORT messages, including the initial PING.

Do a STOP before initial START when probing the device, to try and
reset resources. The subsequent START uses a larger timeout to allow
the firmware to perform all the tasks for the STOP.

INTERNAL: STL-75182

Signed-off-by: Douglas Miller <doug.miller@cornelisnetworks.com>
Signed-off-by: Dennis Dalessandro <dennis.dalessandro@cornelisnetworks.com>