Fix incorrect backport line in non-compiled code
The backported code for fault.c included a typo using ppd instead of
ppr. This does not trigger an error on most distros because fault
injection is not compiled in. It must be on SLES so we want to remove it
and might as well apply the patch to all branches for good measure.

Signed-off-by: Dennis Dalessandro <dennis.dalessandro@cornelisnetworks.com>