RDMA/hfi1: dms: check received packets for error
The hardware will flag certain error conditions when processing a
packet, such as ICRC validation failure, and will set the RheValid bit
in the RHF. The exact error condition is then retrieved from the RHE
queue for the packet.

The dms code must drop packets that are in error.

INTERNAL: STL-75826 hfisvc/dms must not process packets that are in error

Signed-off-by: Michael Blocksome <michael.blocksome@cornelisnetworks.com>
Signed-off-by: Mitch Williams <mitch.williams@cornelisnetworks.com>