bulksvc: Use correct DMS key for read request
The DMS key should contain the remote access and client keys.
This is what gets sent to the remote side for processing, it does not
care about the local client key.

When the local and remote client key index become out of sync, this
will cause all future reads to be out of sync and the remote side
will report:
  hfi1_0: Invalid client error. key = 0x0000000000000004 (4)
  hfi1_0: (4787) Client not found for dms_key 17179869184.

INTERNAL: STL-7583

Signed-off-by: Nick Child <nchild@cornelisnetworks.com>
Signed-off-by: Mitch Williams <mitch.williams@cornelisnetworks.com>